requirements.txt
boto3==1.24.59
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
fasttext==0.9.2
great_expectations==0.15.25
joblib==1.1.0
hydra-core==1.2.0
mlflow==1.28.0
pandas==1.4.4
pybind11==2.10.0
python-dotenv==0.20.0
requests==2.28.1
s3fs==2022.8.2
scikit-learn==1.1.2
scipy==1.9.1
spacy==3.4.1
imbalanced-learn==0.9.1
imblearn==0.0